requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
chromadb==0.3.21
pymupdf==1.22.5
pytesseract==0.3.10
Pillow==9.5.0
sentence-transformers==2.2.2
transformers==4.34.0
streamlit==1.26.1
requests==2.31.0
openai==0.27.8
